Disabled far/near-plan clipping for data
You can zoom-in 10^17 times before you begin to see some far artifacts.
To put it in perspective, This scaling can go from 1 meter to 10ly or
from the size of the observable universe (~10^11ly) to 0.06AU (only a bit
less than ten times the sun's diameter !).
Near objects will be messed up way before that ! It's only a
far-concerned patch.